requirements.txt
langchain
langchain_core
langchain_google_genai
fastapi
python-dotenv
uvicorn
